Revert "LPS-64353 Move"
This reverts commit f0c409a0d8f78020bfe1cd31135bb0f9727c7d29.
