+++ modules/apps/collaboration/social/social-activities-web/build.gradle
@@ -9,6 +9,6 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided group: "rome", name: "rome", version: "1.0"
 	provided project(":apps:foundation:frontend-taglib:frontend-taglib-web")
-	provided project(":apps:foundation:rss:rss-util")
+	provided project(":apps:foundation:portal:portal-rss")
 	provided project(":apps:foundation:portal:portal-upgrade")
 }
