+++ modules/apps/collaboration/wiki/wiki-service/build.gradle
@@ -20,7 +20,7 @@ dependencies {
 	provided project(":apps:foundation:osgi:osgi-service-tracker-collections")
 	provided project(":apps:foundation:osgi:osgi-util")
 	provided project(":apps:foundation:portal:portal-dao-orm-custom-sql")
-	provided project(":apps:foundation:rss:rss-util")
+	provided project(":apps:foundation:portal:portal-rss")
 	provided project(":apps:foundation:portal:portal-spring-extender")
 	provided project(":apps:foundation:portal:portal-upgrade")
 	provided project(":apps:foundation:xstream:xstream-configurator-api")
