+++ modules/apps/forms-and-workflow/calendar/calendar-service/build.gradle
@@ -18,7 +18,7 @@ dependencies {
 	provided group: "rome", name: "rome", version: "1.0"
 	provided project(":apps:forms-and-workflow:calendar:calendar-api")
 	provided project(":apps:foundation:portal:portal-dao-orm-custom-sql")
-	provided project(":apps:foundation:rss:rss-util")
+	provided project(":apps:foundation:portal:portal-rss")
 	provided project(":apps:foundation:portal:portal-spring-extender")
 	provided project(":apps:foundation:portal:portal-upgrade")
 	provided project(":apps:foundation:xstream:xstream-configurator-api")
