+++ modules/apps/forms-and-workflow/calendar/calendar-service/src/main/java/com/liferay/calendar/util/RSSUtil.java
@@ -27,7 +27,7 @@ import java.text.Format;
 /**
  * @author Bruno Basto
  */
-public class RSSUtil extends com.liferay.rss.util.RSSUtil {
+public class RSSUtil extends com.liferay.portal.rss.RSSUtil {
 
 	public static final long TIME_INTERVAL_DEFAULT = Time.WEEK;
