+++ modules/apps/foundation/portal/portal-rss/build.gradle
@@ -5,3 +5,7 @@ dependencies {
 	provided group: "jdom", name: "jdom", version: "1.0"
 	provided group: "rome", name: "rome", version: "1.0"
 }
+
+liferay {
+	deployDir = file("${liferayHome}/osgi/portal")
+}
