+++ modules/apps/foundation/portal/portal-rss/src/main/java/com/liferay/portal/rss/RSSThreadLocal.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.rss.util;
+package com.liferay.portal.rss;
 
 import com.liferay.portal.kernel.util.AutoResetThreadLocal;
