+++ modules/apps/web-experience/asset/asset-publisher-web/build.gradle
@@ -12,7 +12,7 @@ dependencies {
 	provided project(":apps:forms-and-workflow:dynamic-data-mapping:dynamic-data-mapping-api")
 	provided project(":apps:forms-and-workflow:dynamic-data-mapping:dynamic-data-mapping-taglib")
 	provided project(":apps:foundation:frontend-taglib:frontend-taglib-web")
-	provided project(":apps:foundation:rss:rss-util")
+	provided project(":apps:foundation:portal:portal-rss")
 	provided project(":apps:foundation:portal:portal-upgrade")
 	provided project(":apps:web-experience:export-import:export-import-api")
 	provided project(":apps:web-experience:export-import:export-import-service")
