+++ modules/apps/web-experience/journal/journal-web/build.gradle
@@ -19,7 +19,7 @@ dependencies {
 	provided project(":apps:foundation:frontend-taglib:frontend-taglib-web")
 	provided project(":apps:foundation:osgi:osgi-service-tracker-collections")
 	provided project(":apps:foundation:portal-configuration:portal-configuration-metatype")
-	provided project(":apps:foundation:rss:rss-util")
+	provided project(":apps:foundation:portal:portal-rss")
 	provided project(":apps:foundation:portal:portal-upgrade")
 	provided project(":apps:web-experience:application-list:application-list-api")
 	provided project(":apps:web-experience:journal:journal-api")
