+++ modules/apps/web-experience/journal/journal-web/src/main/java/com/liferay/journal/web/util/JournalRSSUtil.java
@@ -59,7 +59,7 @@ import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.Node;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
 import com.liferay.portal.kernel.xml.XPath;
-import com.liferay.rss.util.RSSUtil;
+import com.liferay.portal.rss.RSSUtil;
 
 import com.sun.syndication.feed.synd.SyndContent;
 import com.sun.syndication.feed.synd.SyndContentImpl;
