+++ modules/apps/web-experience/journal/journal-web/src/main/resources/META-INF/resources/init.jsp
@@ -164,7 +164,7 @@ page import="com.liferay.portal.kernel.workflow.WorkflowDefinitionManagerUtil" %
 page import="com.liferay.portal.kernel.workflow.WorkflowEngineManagerUtil" %><%@
 page import="com.liferay.portal.kernel.workflow.WorkflowHandlerRegistryUtil" %><%@
 page import="com.liferay.portal.model.impl.*" %><%@
-page import="com.liferay.rss.util.RSSUtil" %><%@
+page import="com.liferay.portal.rss.RSSUtil" %><%@
 page import="com.liferay.portal.service.*" %><%@
 page import="com.liferay.portal.upload.LiferayFileItem" %><%@
 page import="com.liferay.portal.util.PrefsPropsUtil" %><%@
